refactor: move constants to test
